# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "board Rest API",
        "description": "board API",
        "contact": {
            "email": "chkim@dtwocorp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": ""
        }
    ],
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "/users"
                ],
                "summary": "유저 목록 가져오기",
                "description": "제공된 필터와 일치하는 모든 사용자 또는 사용자의 리턴 목록",
                "operationId": "12e989554c281d650dff75443acabf54",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "이름 필터",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Users not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "/users"
                ],
                "summary": "새로운 유저 생성",
                "description": "새로운 유저를 생성하고 사용자 데이터를 반환합니다.",
                "operationId": "64da9b5a20fa4f20e947ed61b2bbb6da",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/users/{email}": {
            "get": {
                "tags": [
                    "/users"
                ],
                "summary": "이메일로 유저 정보 가져오기",
                "description": "유저 데이터를 이메일로 가져오고 리턴합니다.",
                "operationId": "bae3447980ccbdcd0f7471133e4867cd",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "유저의 이메일",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "/users"
                ],
                "summary": "유저 삭제하기",
                "description": "제공받은 이메일로 유저를 삭제합니다.",
                "operationId": "b6cc2e6a8ed518dcc0c4007b82ef7f40",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "The email of the user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "User deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "/users"
                ],
                "summary": "이메일로 유저 정보 업데이트하기",
                "description": "제공받은 이메일로 유저 정보를 업데이트하고 리턴합니다.",
                "operationId": "f2891a3c873f68cf99a5da2fe3b944cc",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "유저의 이메일",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    }
                ],
                "requestBody": {
                    "description": "새로운 유저 데이터",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserResource"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "유저 정보 업데이트 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserResource": {
                "properties": {
                    "Name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "Created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}